requirements/dev.txt
-r base.txt
django-debug-toolbar
ipdb
factory_boy
fake-factory